refactor: improve error handling and consolidate tests
- Validate and reject whitespace-only data source values

- Narrow exception handling in serialization helpers (AttributeError, TypeError, ValueError, RuntimeError)

- Consolidate model test files by merging test_model_class.py into test_models.py

- Remove duplicate test_accepts_string_path test

Made-with: Cursor
