+++ src/raitap/transparency/results.py
@@ -28,7 +28,7 @@ def _serialisable(value: Any) -> Any:
     if hasattr(value, "item"):
         try:
             return _serialisable(value.item())
-        except Exception:
+        except (AttributeError, TypeError, ValueError, RuntimeError):
             pass
     return repr(value)
